Promote high-level structs to main namespace
